make the folder structure consistent
